subcategorizeOne
